pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lerim"
version = "0.1.0"
description = "Continual learning layer for coding agents and software projects."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSL-1.1" }
dependencies = [
    "pydantic>=2.0",
    "eval_type_backport>=0.2.0; python_version < '3.10'",
    "tomli>=2.0; python_version < '3.11'",
    "python-dotenv>=1.2.1",
    "PyYAML>=6.0",
    "loguru>=0.7.3",
    "lancedb>=0.27.1",
    "pyarrow>=15.0.0",
    "torch>=2.2",
    "transformers>=4.44.2,<5",
    "dspy>=3.1.3",
    "python-frontmatter>=1.1.0",
    "openrouter>=0.6.0",
    "pydantic-ai>=1.61.0",
    "logfire>=4.25.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.15.0",
]

[project.scripts]
lerim = "lerim.app.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["lerim*"]

[tool.setuptools.package-data]
lerim = ["config/default.toml"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that hit real external services",
    "e2e: marks end-to-end tests that exercise full flows",
    "agent: agent SDK or CLI integration tests",
    "llm: non-agentic LLM integration tests",
    "embeddings: embedding integration tests",
    "openrouter: OpenRouter provider tests",
    "openai: OpenAI provider tests",
    "zai: ZAI provider tests",
    "kimi: Kimi provider tests",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
extend-select = ["B006", "T201"]
ignore = ["E501"]
